+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Created `util` package to contain `HashUtil` and `ProgramHandlingUtil`
+- All input files are unpacked to a temporary directory.
+
 ## [1.0.7] - 2022-02-10
 
 ### Added